+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Unreleased
+* **Clarisse submitter:**
+  * Fixed bug where render file would be cleaned up before the upload daemon had a chance to upload it.
 
+# v2.11.1
 * **Clarisse submitter:**
   * Catch invalid glob path that caused Clarisse to crash.